Remove nesting for JoinTable#joinColumns and Table#indexes/uniqueConstraints again.
